+++ app/subscriber/src/features/search-page/components/advanced-search/components/sections/MediaSection.tsx
@@ -1,3 +1,4 @@
+import { createSubMediaGroups } from 'features/utils/createSubMediaGroups';
 import React from 'react';
 import { IoIosArrowDropdownCircle, IoIosArrowDroprightCircle } from 'react-icons/io';
 import { useContent } from 'store/hooks';
@@ -32,44 +33,9 @@ export const MediaSection: React.FC<IMediaSectionProps> = ({
   const [mediaGroupExpandedStates, setMediaGroupExpandedStates] =
     React.useState<ISubMediaGroupExpanded>();
   const [subMediaGroups, setSubMediaGroups] = React.useState<ISubMediaGroupItem[]>();
+
   React.useEffect(() => {
-    // exit early if inputs are not set completely
-    if (sources.length === 0 || mediaTypes.length === 0) return;
-
-    let mediaTypeSourceLookup: { [name: string]: ISourceModel[] } = {};
-    const allSourcesKey: string = 'All';
-    mediaTypeSourceLookup[allSourcesKey] = [];
-    // prime the dictionary - already in sort order set on Media Type
-    mediaTypes.forEach((mt) => {
-      mediaTypeSourceLookup[mt.name] = [];
-    });
-    sources.forEach((source) => {
-      mediaTypeSourceLookup[allSourcesKey].push(source);
-      source.mediaTypeSearchMappings.forEach((mapping) => {
-        mediaTypeSourceLookup[mapping.name].push(source);
-      });
-    });
-    // Remove Media Type entries with no assigned Sources
-    // Could also exclude specific Media Types her if required
-    mediaTypeSourceLookup = Object.fromEntries(
-      Object.entries(mediaTypeSourceLookup).filter(([k, v]) => v.length > 0),
-    );
-
-    let subMediaGroups: ISubMediaGroupItem[] = [];
-    let mediaGroupExpandedStates: ISubMediaGroupExpanded = {};
-    for (let key in mediaTypeSourceLookup) {
-      // Use `key` and `value`
-      let value = mediaTypeSourceLookup[key];
-      subMediaGroups.push({
-        key: key,
-        label: key,
-        options: value,
-      });
-      // initialize state model for what groups are expanded/collapsed
-      mediaGroupExpandedStates[key] = false;
-    }
-    setSubMediaGroups(subMediaGroups);
-    setMediaGroupExpandedStates(mediaGroupExpandedStates);
+    createSubMediaGroups(sources, mediaTypes, setSubMediaGroups, setMediaGroupExpandedStates);
   }, [sources, mediaTypes, setSubMediaGroups, setMediaGroupExpandedStates]);
 
   const [
